build: move ExampleNodeWallet to own entrypoint
No need to pollute the global API or bundle size with this.
